Reduce duplicate function calls, add debug messages
I figured it would be more likely a user would want to fall back to default behavior of new graves not being made if the oldest couldn't be broken. Otherwise, you might run into infinite graves being made while the issue isn't fixed (I realized having an incorrect system time is one potential cause for this issue).

I also made the getOldestGrave() method take in a list of graves as an argument instead, so we only need to call for the list creation once in EntityDeathListener class. This probably would hurt the method's potential usage in other classes, but tbh I really don't see it being used anywhere else. This is all just for readability's sake